move input components to sub folder
